Backend/requirements.txt
flask
flask_cors
gunicorn
pyjwt
werkzeug
psycopg2-binary